Moved all components into their own folders
